oldland: fix varags passing for more than 6 arguments.
The first 6 arguments are passed in registers, the rest need to be
passed by stack.  This works by pushing all register parameters onto the
stack such that they are contiguous with the stack parameters in the
caller.

With this change we don't need the caller to reserve space for registers
that would normally be passed, instead we manually fix the frame and
stack pointers during the prologue computation.